+++ src/affindaAPI.ts
@@ -1294,7 +1294,12 @@ const listOccupationGroupsOperationSpec: coreClient.OperationSpec = {
   httpMethod: "GET",
   responses: {
     201: {
-      bodyMapper: Mappers.OccupationGroup
+      bodyMapper: {
+        type: {
+          name: "Sequence",
+          element: { type: { name: "Composite", className: "OccupationGroup" } }
+        }
+      }
     },
     400: {
       bodyMapper: Mappers.RequestError